Improve handling of children props
Add DOCGEN_Children to allow adding children prop manually where it is not in docs json - add to CardContents.

DocGenContext now passes through children props - they are not removed by useProp. This allows DocGen to base decision on supporting children on the processed component rather than the raw one, so inheriting children props now works. They are filtered in DocGen instead, when producing the usedProps to actually generate code from directly.

Move sanitisation of prop descriptions from DocGenContext to DocGen, since it is a function of code generation not the individual React library, and so shouldn't change. This also allows `@ignore` comments to pass through and be used to determine whether children props should be used.